+++ TDEngine2/include/core/IGraphicsContext.h
@@ -378,6 +378,8 @@ namespace TDEngine2
 
 			TDE2_API virtual const TGraphicsContextInfo& GetContextInfo() const = 0;
 
+			TDE2_API virtual std::vector<U8> GetBackBufferData() const = 0;
+
 			/*!
 				\return The method returns a pointer to IWindowSystem
 			*/